feat: (#73) 로그인 상태체크 및 처리 로직 이동함 (메인->app/로그인페이지)
