+++ frontend/.env
@@ -1,4 +1,4 @@
 SKIP_PREFLIGHT_CHECK=true
 REACT_APP_NODE_PATH=./src
 REACT_APP_DB_BASIC_URL=http://3.37.76.224/api 
-REACT_APP_MAIN_URL=http://localhost:3000/main
+REACT_APP_BASIC_URL=http://localhost:3000/
