+++ frontend/src/page/mainPage/MainPage.tsx
@@ -1,38 +1,14 @@
-import React, { Suspense, useEffect } from 'react';
+import React, { Suspense } from 'react';
 import styled from 'styled-components';
 import IssueTable from 'page/mainPage/issueTable/IssueTable';
 import OptionTable from 'page/mainPage/optionTable/OptionTable';
-import { fetchLogin } from 'util/api/fetchLogin';
-import { useRecoilState, useSetRecoilState, useResetRecoilState } from 'recoil';
-import { controlLoginState } from 'store/loginStore';
-import { getIssueTrigger } from 'store/issueInfoStore';
+import { useResetRecoilState } from 'recoil';
+
 import { resetTabClickedState } from 'store/labelMilestoneStore';
 
 export default function MainPage() {
   const resetTabClicked = useResetRecoilState(resetTabClickedState);
   resetTabClicked();
-  const setLogin = useSetRecoilState(controlLoginState);
-  const [issueTrigger, setIssueTrigger] = useRecoilState(getIssueTrigger);
-
-  useEffect(() => {
-    if (isLogin()) return;
-    const query = window.location.search;
-    const loginCode = query.split('=')[1];
-    setLoginData(loginCode);
-    setIssueTrigger(false);
-  }, []);
-
-  const setLoginData = async (loginCode: string) => {
-    try {
-      const loginData = await fetchLogin(loginCode);
-      setLogin(loginData);
-      localStorage.setItem('token', loginData.token);
-    } catch (err) {
-      throw err;
-    }
-  };
-
-  const isLogin = () => !!localStorage.getItem('token');
 
   return (
     <MainPageBlock>
